Move FFPlayState and FFPlayStateType into play.state package
